Move aes tests to their own file
